moved toolbar expand/collapse methods to CollapsingToolbarState
